refactor into a separate container-worker crate
