+++ workers/rust/Cargo.toml
@@ -1,5 +1,6 @@
 [workspace]
 
 members = [
+    "container",
     "lib-worker",
 ]
