+++ workers/rust/container/src/bin/docker-worker.rs
@@ -1,0 +1,12 @@
+use slog::{o, Drain, Logger};
+
+#[tokio::main]
+async fn main() {
+    let decorator = slog_term::TermDecorator::new().build();
+    let drain = slog_term::FullFormat::new(decorator).build().fuse();
+    let drain = slog_async::Async::new(drain).build().fuse();
+
+    let logger = Logger::root(drain, o!());
+
+    container_worker::main(logger).await;
+}
